global: add pre_filters for repeated sources
This adds pre_filters to the merger rules that can preprocess the
records before handing them over to the merger, and implements a
pre-filter to remove the documents and figures from the head and root
matching the source in the update.

It also renames `filter_out` to `filter_conflicts` or
`conflict_filters` (for the function and attributes respectively) and
simplifies `is_to_delete`.

Signed-off-by: Micha Moskovic <michamos@gmail.com>